backend/requirements.txt
fastapi==0.136.0
uvicorn==0.46.0
sqlalchemy==2.0.49
psycopg2==2.9.12
pydantic==2.13.3
python-dotenv==1.2.2
sqlmodel==-0.0.38